requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# Bug-Squasher-Bot/gdrive.py: 2,3
google_api_python_client == 2.4.0

# Bug-Squasher-Bot/gdrive.py: 1
protobuf == 3.12.4

# Bug-Squasher-Bot/main.py: 4
python_telegram_bot == 13.5
